+++ demos/react-native-supabase-todolist/library/widgets/CameraWidget.tsx
@@ -1,4 +1,4 @@
-import { Camera, CameraCapturedPicture, CameraType } from 'expo-camera';
+import { CameraView, useCameraPermissions, CameraCapturedPicture } from 'expo-camera';
 import React, { useRef } from 'react';
 import {
   StyleSheet,
@@ -9,7 +9,9 @@ import {
   ActivityIndicator,
   Platform
 } from 'react-native';
-import { Button, Icon } from 'react-native-elements';
+import { Button, Icon } from '@rneui/themed';
+import { SafeAreaView } from 'react-native-safe-area-context';
+import { CameraType } from 'expo-camera/build/legacy/Camera.types';
 
 export interface Props {
   onCaptured: (photo: CameraCapturedPicture) => void;
@@ -19,10 +21,11 @@ export interface Props {
 const isAndroid = Platform.OS === 'android';
 
 export const CameraWidget: React.FC<Props> = (props) => {
-  const cameraRef = useRef<Camera>(null);
-  const [permission, requestPermission] = Camera.useCameraPermissions();
+  const cameraRef = useRef<CameraView>(null);
+  const [permission, requestPermission] = useCameraPermissions();
   const [ready, setReady] = React.useState(false);
   const [loading, setLoading] = React.useState(false);
+  const [facing, setFacing] = React.useState(CameraType.back);
 
   // Getting width and height of the screen
   const { width } = useWindowDimensions();
@@ -41,7 +44,7 @@ export const CameraWidget: React.FC<Props> = (props) => {
       };
       const photo = await cameraRef.current.takePictureAsync(options);
       setLoading(false);
-      props.onCaptured(photo);
+      if (photo) props.onCaptured(photo);
       props.onClose();
     }
   };
@@ -52,6 +55,10 @@ export const CameraWidget: React.FC<Props> = (props) => {
     } else {
       setReady(true);
     }
+  };
+
+  const onFlipPress = () => {
+    setFacing(facing === CameraType.back ? CameraType.front : CameraType.back);
   };
 
   if (!permission) {
@@ -70,48 +77,71 @@ export const CameraWidget: React.FC<Props> = (props) => {
   }
 
   return (
-    <View style={styles.container}>
-      <Camera
+    <SafeAreaView style={styles.container}>
+      <View style={styles.topBarContainer}>
+        <TouchableOpacity onPress={props.onClose} style={styles.topBarCloseIcon}>
+          <Icon name={'close'} type={'material'} color={'white'} size={32} />
+        </TouchableOpacity>
+
+        <TouchableOpacity onPress={onFlipPress} style={styles.topBarFlipIcon}>
+          <Icon name={'flip-camera-ios'} type={'material'} color={'white'} size={32} />
+        </TouchableOpacity>
+      </View>
+      <CameraView
         ref={cameraRef}
         style={{ ...styles.camera, height: height, width: '100%' }}
-        type={CameraType.back}
-        onCameraReady={onReady}
-        ratio={isAndroid ? '16:9' : undefined}
-      >
-        <TouchableOpacity onPress={props.onClose} style={styles.backButton}>
-          <Icon name={'chevron-left'} type={'font-awesome'} color={'white'} />
+        facing={facing}
+        onCameraReady={onReady}></CameraView>
+      <View style={styles.bottomCamera}>
+        <TouchableOpacity disabled={loading} style={styles.shutterButton} onPress={captureImageAsync}>
+          <ActivityIndicator animating={loading} />
         </TouchableOpacity>
-        <View style={styles.bottomCamera}>
-          <TouchableOpacity disabled={loading} style={styles.shutterButton} onPress={captureImageAsync}>
-            <ActivityIndicator animating={loading} />
-          </TouchableOpacity>
-        </View>
-      </Camera>
-    </View>
+      </View>
+    </SafeAreaView>
   );
 };
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    justifyContent: 'center'
+    flexDirection: 'column',
+    justifyContent: 'center',
+    backgroundColor: 'transparent'
   },
-  backButton: {
-    position: 'absolute',
-    top: 16,
-    left: 16
+  topBarContainer: {
+    flex: 0,
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    padding: 20,
+    width: '100%',
+    backgroundColor: 'black',
+    height: 80
+  },
+  topBarCloseIcon: {
+    flex: 1,
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'flex-start'
+  },
+  topBarFlipIcon: {
+    flex: 1,
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'flex-end'
   },
   camera: {
     flex: 1
   },
   bottomCamera: {
-    flex: 1,
-    alignSelf: 'stretch',
+    flex: 0,
+    // alignSelf: 'stretch',
     alignItems: 'flex-end',
     justifyContent: 'center',
     flexDirection: 'row',
     columnGap: 30,
-    backgroundColor: 'transparent'
+    backgroundColor: 'black',
+    minHeight: 100
   },
   shutterButton: {
     width: 70,