+++ demos/react-native-supabase-todolist/library/widgets/HeaderWidget.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { Alert, Text } from 'react-native';
-import { Icon } from 'react-native-elements';
 import { useNavigation } from 'expo-router';
-import { Header } from 'react-native-elements';
+import { Icon, Header } from '@rneui/themed';
 import { useStatus } from '@powersync/react';
 import { DrawerActions } from '@react-navigation/native';
 import { useSystem } from '../powersync/system';
@@ -33,7 +32,7 @@ export const HeaderWidget: React.FC<{
         <Icon
           name={status.connected ? 'wifi' : 'wifi-off'}
           type="material-community"
-          color="black"
+          color="white"
           size={20}
           style={{ padding: 5 }}
           onPress={() => {
@@ -49,7 +48,8 @@ export const HeaderWidget: React.FC<{
           }}
         />
       }
-      centerComponent={<Text style={{ padding: 5, color: '#fff' }}>{title}</Text>}
+      centerContainerStyle={{ justifyContent: 'center', alignItems: 'center' }}
+      centerComponent={{ text: title, style: { color: '#fff' } }}
     />
   );
 };