+++ demos/react-native-supabase-todolist/library/widgets/TodoItemWidget.tsx
@@ -1,12 +1,13 @@
 import { CameraCapturedPicture } from 'expo-camera';
 import React from 'react';
 import { ActivityIndicator, Alert, View, Modal, StyleSheet } from 'react-native';
-import { ListItem, Button, Icon, Image } from 'react-native-elements';
+import { ListItem, Button, Icon, Image } from '@rneui/themed';
 import { CameraWidget } from './CameraWidget';
 import { TodoRecord } from '../powersync/AppSchema';
 import { AttachmentRecord } from '@powersync/attachments';
 import { AppConfig } from '../supabase/AppConfig';
 import { useSystem } from '../powersync/system';
+import { SafeAreaProvider } from 'react-native-safe-area-context';
 
 export interface TodoItemWidgetProps {
   record: TodoRecord;
@@ -29,19 +30,31 @@ export const TodoItemWidget: React.FC<TodoItemWidgetProps> = (props) => {
   return (
     <View key={`todo-item-${record.id}`} style={{ padding: 10 }}>
       <Modal animationType="slide" transparent={false} visible={isCameraVisible} onRequestClose={handleCancel}>
-        <CameraWidget onCaptured={onSavePhoto} onClose={handleCancel} />
+        <SafeAreaProvider>
+          <CameraWidget onCaptured={onSavePhoto} onClose={handleCancel} />
+        </SafeAreaProvider>
       </Modal>
       <ListItem.Swipeable
         bottomDivider
         rightContent={
           <Button
+            buttonStyle={{
+              flexDirection: 'column',
+              alignContent: 'center',
+              borderColor: 'transparent',
+              minHeight: '100%'
+            }}
             containerStyle={{
               flex: 1,
+              flexGrow: 1,
+              height: '100%',
               justifyContent: 'center',
-              backgroundColor: '#d3d3d3'
+              backgroundColor: 'rgba(240, 15, 15, 0.9)'
             }}
+            title="Delete"
+            titleStyle={{ color: 'white' }}
             type="clear"
-            icon={{ name: 'delete', color: 'red' }}
+            icon={{ name: 'trash', type: 'font-awesome', color: 'white' }}
             onPress={() => {
               Alert.alert(
                 'Confirm',
@@ -51,8 +64,7 @@ export const TodoItemWidget: React.FC<TodoItemWidgetProps> = (props) => {
               );
             }}
           />
-        }
-      >
+        }>
         {loading ? (
           <ActivityIndicator />
         ) : (
@@ -69,14 +81,25 @@ export const TodoItemWidget: React.FC<TodoItemWidgetProps> = (props) => {
           />
         )}
         <ListItem.Content style={{ minHeight: 80 }}>
-          <ListItem.Title>{record.description}</ListItem.Title>
+          <ListItem.Title style={{ fontSize: 20 }}>{record.description}</ListItem.Title>
         </ListItem.Content>
         {AppConfig.supabaseBucket &&
           (record.photo_id == null ? (
-            <Icon name={'camera'} type="font-awesome" onPress={() => setCameraVisible(true)} />
+            <Button
+              type="outline"
+              buttonStyle={{ borderColor: 'transparent' }}
+              onPress={() => setCameraVisible(true)}
+              icon={{
+                name: 'camera',
+                type: 'font-awesome',
+                color: 'black',
+                size: 28
+              }}>
+              {/* <Icon name={'camera'} type="material" color={'black'} size={32} /> */}
+            </Button>
           ) : photoAttachment?.local_uri != null ? (
             <Image
-              source={{ uri: system.attachmentQueue.getLocalUri(photoAttachment.local_uri) }}
+              source={{ uri: system.attachmentQueue?.getLocalUri(photoAttachment.local_uri) }}
               containerStyle={styles.item}
               PlaceholderContent={<ActivityIndicator />}
             />