+++ tools/diagnostics-app/src/library/powersync/ConnectionManager.ts
@@ -10,9 +10,18 @@ import Logger from 'js-logger';
 import { DynamicSchemaManager } from './DynamicSchemaManager';
 import { RecordingStorageAdapter } from './RecordingStorageAdapter';
 import { TokenConnector } from './TokenConnector';
+import { safeParse } from '../safeParse/safeParse';
 
 Logger.useDefaults();
 Logger.setLevel(Logger.DEBUG);
+
+export const PARAMS_STORE = 'currentParams';
+
+export const getParams = () => {
+  const stringifiedParams = localStorage.getItem(PARAMS_STORE);
+  const params = safeParse(stringifiedParams);
+  return params;
+}
 
 export const schemaManager = new DynamicSchemaManager();
 
@@ -23,7 +32,7 @@ export const db = new PowerSyncDatabase({
   },
   schema: schemaManager.buildSchema()
 });
-db.execute('PRAGMA cache_size=-50000');
+db.execute('PRAGMA cache_size=-500000');
 
 export const connector = new TokenConnector();
 
@@ -74,7 +83,8 @@ if (connector.hasCredentials()) {
 }
 
 export async function connect() {
-  await sync.connect();
+  const params = getParams();
+  await sync.connect({ params });
   if (!sync.syncStatus.connected) {
     // Disconnect but don't wait for it
     sync.disconnect();
@@ -90,7 +100,8 @@ export async function clearData() {
   await schemaManager.clear();
   await schemaManager.refreshSchema(db.database);
   if (connector.hasCredentials()) {
-    await sync.connect();
+    const params = getParams();
+    await sync.connect({ params });
   }
 }
 
@@ -102,5 +113,11 @@ export async function signOut() {
   connector.clearCredentials();
   await db.disconnectAndClear();
 }
+
+export const setParams = (p: object) => {
+  const stringified = JSON.stringify(p);
+  localStorage.setItem(PARAMS_STORE, stringified);
+  connect();
+};
 
 (window as any).db = db;