+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @powersync/react-native
 
+## 1.14.1
+
+### Patch Changes
+
+- a3f625e: Fix inline require for react-native-quick-sqlite
+
+## 1.14.0
+
+### Minor Changes
+
+- 77e196d: Use powersync-sqlite-core 0.3.0 - faster incremental sync
+
+### Patch Changes
+
+- Updated dependencies [77e196d]
+  - @powersync/common@1.20.0
+
+## 1.13.0
+
+### Minor Changes
+
+- 9c140b5: Make `react-native-quick-sqlite` an optional dependency so that it can be used conditionally.
+
+## 1.12.4
+
+### Patch Changes
+
+- Updated dependencies [f8ac369]
+  - @powersync/react@1.4.5
+
 ## 1.12.3
 
 ### Patch Changes